requirements.txt
# --- Core UI ---
streamlit==1.39.0
python-dotenv==1.0.1

# --- LangChain & Gemini Integration ---
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.3
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

# --- Database & ORM ---
SQLAlchemy==2.0.36

# --- Supporting Libraries ---
protobuf==5.26.1
pandas==2.2.2
matplotlib==3.9.2

# --- Optional for advanced visualization ---
plotly==5.24.1

# --- Utility Tools ---
tqdm==4.66.4
requests==2.32.3
